feat(@meso-network/meso-js): EOL @meso-network/types in favor of collapsing into @meso-network/meso-js
