+++ .vscode/settings.json
@@ -1,9 +1,5 @@
 {
-  "conventionalCommits.scopes": [
-    "@meso-network",
-    "@meso-network/meso-js",
-    "@meso-network/types"
-  ],
+  "conventionalCommits.scopes": ["@meso-network", "@meso-network/meso-js"],
   "editor.formatOnSave": true,
   "editor.defaultFormatter": "esbenp.prettier-vscode",
   // https://github.com/microsoft/vscode-eslint/issues/1620#issuecomment-1473305698
