+++ packages/meso-js/src/index.ts
@@ -1,6 +1,6 @@
 import { version } from "../package.json";
 
-export * from "@meso-network/types";
+export * from "./types";
 export { transfer } from "./transfer";
 export { validateLayout } from "./validateLayout";
 export const MESO_JS_VERSION = version;
