--- packages/types/README.md
@@ -1,7 +1,0 @@
-# `@meso-network/types`
-
-This package provides common TypeScript definitions for [`meso-js`](../meso-js/).
-
-**Why?** The nature of the Meso experience requires these types to be shared
-by our public libraries as well as internal products (such as the iframe that
-hosts the Meso on/off-ramp).
