--- packages/types/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "@meso-network/types",
-  "version": "0.0.71",
-  "description": "Common TypeScript definitions for @meso-network packages.",
-  "main": "src/index.ts",
-  "types": "src/index.ts",
-  "private": true,
-  "scripts": {},
-  "keywords": [],
-  "author": "",
-  "license": "MIT"
-}
