--- packages/types/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "rootDir": "./",
-    "composite": true,
-    "declaration": true,
-    "emitDeclarationOnly": true
-  },
-  "include": ["src"]
-}
